fixed tests after minor changes

+++ dev-scripts/run-tests.sh
@@ -63,7 +63,13 @@ if [ $TESTTYPE == "phpunit" ]; then
   INI_PATH=/etc/php/8.2/cli/conf.d/99-phpunit.ini
   cecho b "# Php Unit Tests"
   docker $DOCKER_EXECPARAMS "printf 'zend_extension=xdebug.so\nxdebug.mode=coverage\nmemory_limit=-1' > $INI_PATH"
-  docker $DOCKER_EXECPARAMS "mkdir -p /framelix/userdata/tmp && chmod 0777 -R /framelix/userdata/tmp && cd /framelix/appdata && composer update && framelix_php vendor/bin/phpunit --coverage-clover /framelix/userdata/tmp/clover.xml --bootstrap modules/FramelixTests/tests/_bootstrap.php --configuration  modules/FramelixTests/tests/_phpunit.xml -d zend_extension=xdebug.so -d xdebug.mode=coverage -d memory_limit=-1 && framelix_php hooks/after-phpunit.php"
+  if [ "$?" != "0" ]; then
+    exit $?
+  fi
+  docker $DOCKER_EXECPARAMS "mkdir -p /framelix/userdata/tmp && chmod 0777 -R /framelix/userdata/tmp && cd /framelix/appdata && composer update && framelix_php vendor/bin/phpunit --coverage-clover /framelix/userdata/tmp/clover.xml --bootstrap modules/FramelixTests/tests/_bootstrap.php --configuration modules/FramelixTests/tests/_phpunit.xml && framelix_php hooks/after-phpunit.php"
+  if [ "$?" != "0" ]; then
+    exit $?
+  fi
   docker $DOCKER_EXECPARAMS "rm -f $INI_PATH"
   exit $?
 fi
